+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Outsystems.SetupTools Release History
 
+## 2.6.0.0
+
+### Changes
+
+- New function Write-OSInstallLog. To write messages on the install log and the verbose stream
+- Install-OSServer. Added -FullPathInstallDir switch
+- Changed FullPathInstallDir parameter on Install-OSServer and Install-OSServiceStudio to dynamic. Only available when InstallDir parameter is specified
+
 ## 2.5.3.0
 
 ### Changes